* bitbucket/master:
  version info: update
  ChangeLog: update
  size_t --> int

+++ src/NetBox/FarPlugin.cpp
@@ -2763,7 +2763,7 @@ void __fastcall TFarPanelInfo::ApplySelection()
 //---------------------------------------------------------------------------
 TFarPanelItem * __fastcall TFarPanelInfo::GetFocusedItem()
 {
-  size_t Index = GetFocusedIndex();
+  int Index = GetFocusedIndex();
   TObjectList * Items = GetItems();
   // DEBUG_PRINTF(L"Index = %d, Items = %x, Items->GetCount = %d", Index, Items, Items->GetCount());
   if (Items->GetCount() > 0)
@@ -2780,7 +2780,7 @@ TFarPanelItem * __fastcall TFarPanelInfo::GetFocusedItem()
 void __fastcall TFarPanelInfo::SetFocusedItem(TFarPanelItem * value)
 {
   TObjectList * Items = GetItems();
-  size_t Index = Items->IndexOf(static_cast<TObject *>(value));
+  int Index = Items->IndexOf(static_cast<TObject *>(value));
   assert(Index != NPOS);
   SetFocusedIndex(Index);
   // delete Items;
@@ -2788,7 +2788,7 @@ void __fastcall TFarPanelInfo::SetFocusedItem(TFarPanelItem * value)
 //---------------------------------------------------------------------------
 int __fastcall TFarPanelInfo::GetFocusedIndex()
 {
-  return static_cast<size_t>(FPanelInfo->CurrentItem);
+  return FPanelInfo->CurrentItem;
 }
 //---------------------------------------------------------------------------
 void __fastcall TFarPanelInfo::SetFocusedIndex(int value)
